Give the board wiring and the battery verdict real types
Pin numbers were macros, against a standard that asks for constexpr or
enum class so the values carry a type and cannot be quietly mixed with
numbers that mean something else. Nothing used them in a preprocessor
condition and no C file includes the header, so they convert cleanly. The
SPI and I2C instances stay macros -- spi1 and i2c1 are the SDK's register
addresses cast to pointers, which constexpr cannot hold -- as does the SDIO
switch, which is meant for an #if.

battery_presence() returned 1 or 2 as bare ints, and had stopped being able
to return 0 when the deeply discharged cell was reclassified. Saying that in
the type is better than saying it in a comment: BatteryPresence has no value
for absence, so the guarantee holds by construction rather than by everyone
remembering. The two numbers are pinned to the enumerators because BATTERY(2)
hands them to BASIC programs and MANUAL.md documents them.

The 4100mV threshold picks up a name while it is in reach.

+++ src/expression_parser.cpp
@@ -208,7 +208,9 @@ static Value evaluate_builtin_function(const char* var_name, Value* args, int ar
         switch ((int)args[0].num_val) {
             case 0: return Value(mv);
             case 1: return Value(battery_percent_from_mv(mv));
-            case 2: return Value(battery_presence(mv, hal_battery_usb_connected()));
+            // BATTERY(2) は 1=有り / 2=判別不能。数は BatteryPresence の値と対
+            case 2: return Value(static_cast<int>(
+                        battery_presence(mv, hal_battery_usb_connected())));
             case 3:
                 // USB から給電されているか。VBUS を読む線が無いので
                 // USB シリアルの接続で代用している（充電器だけの接続は 0）
